Moves testing of valid schema cache to the end
This way all schemas in the tests suite are considered.
